requirements.txt
requests
bs4
pd2image
pdf2image
pdf2image
pillow
pandas
flask
flask-restful
flask-cors
textract
pdf2image
textract
pymongo